♻️ Refactor authenticity metadata
